+++ crates/zng-wgt-image/src/mask.rs
@@ -3,7 +3,7 @@
 //! [`mask_image`]: fn@mask_image
 //! [`mask_mode`]: fn@mask_mode
 
-use zng_ext_image::{IMAGES, ImageCacheMode, ImageDownscale, ImageLimits, ImageMaskMode, ImageRenderArgs, ImageSource};
+use zng_ext_image::{IMAGES, ImageCacheMode, ImageDownscale, ImageEntriesMode, ImageLimits, ImageMaskMode, ImageRenderArgs, ImageSource};
 use zng_wgt::prelude::*;
 
 use crate::ImageFit;
@@ -44,7 +44,7 @@ pub fn mask_image(child: impl IntoUiNode, source: impl IntoVar<ImageSource>) -> 
             if let ImageSource::Render(_, args) = &mut source {
                 *args = Some(ImageRenderArgs::new(WINDOW.id()));
             }
-            let i = IMAGES.image(source, mode, limits, downscale, Some(mask_mode));
+            let i = IMAGES.image2(source, mode, limits, downscale, Some(mask_mode), ImageEntriesMode::PRIMARY);
             let s = i.subscribe(UpdateOp::Update, WIDGET.id());
             img = Some((i, s));
 
@@ -77,7 +77,7 @@ pub fn mask_image(child: impl IntoUiNode, source: impl IntoVar<ImageSource>) -> 
                 let downscale = MASK_IMAGE_DOWNSCALE_VAR.get();
                 let mask_mode = MASK_MODE_VAR.get();
 
-                let i = IMAGES.image(source, mode, limits, downscale, Some(mask_mode));
+                let i = IMAGES.image2(source, mode, limits, downscale, Some(mask_mode), ImageEntriesMode::PRIMARY);
                 let s = i.subscribe(UpdateOp::Layout, WIDGET.id());
                 img = Some((i, s));
 
@@ -98,7 +98,14 @@ pub fn mask_image(child: impl IntoUiNode, source: impl IntoVar<ImageSource>) -> 
                         let limits = MASK_IMAGE_LIMITS_VAR.get();
                         let downscale = MASK_IMAGE_DOWNSCALE_VAR.get();
                         let mask_mode = MASK_MODE_VAR.get();
-                        IMAGES.image(source, ImageCacheMode::Cache, limits, downscale, Some(mask_mode))
+                        IMAGES.image2(
+                            source,
+                            ImageCacheMode::Cache,
+                            limits,
+                            downscale,
+                            Some(mask_mode),
+                            ImageEntriesMode::PRIMARY,
+                        )
                     };
 
                     let s = i.subscribe(UpdateOp::Update, WIDGET.id());