+++ crates/zng/src/image.rs
@@ -72,8 +72,8 @@
 //! See [`zng_ext_image`] for the full image API and [`zng_wgt_image`] for the full widget API.
 
 pub use zng_ext_image::{
-    IMAGE_RENDER, IMAGES, ImageCacheMode, ImageDataFormat, ImageDownscale, ImageHash, ImageHasher, ImageLimits, ImageRenderArgs,
-    ImageSource, ImageSourceFilter, ImageVar, Img, PathFilter, render_retain,
+    IMAGE_RENDER, IMAGES, ImageCacheMode, ImageDataFormat, ImageDownscale, ImageEntriesMode, ImageEntryKind, ImageHash, ImageHasher,
+    ImageLimits, ImageRenderArgs, ImageSource, ImageSourceFilter, ImageVar, Img, PathFilter, render_retain,
 };
 
 #[cfg(feature = "http")]